# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Supercheck API",
    "version": "1.0.0",
    "description": "API for test automation, monitoring, and CI/CD integration. Supercheck provides Playwright browser testing, k6 performance testing, uptime monitoring, and alerting — all managed via API or CLI.",
    "contact": {
      "name": "Supercheck",
      "url": "https://supercheck.io"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://github.com/supercheck-io/supercheck/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://app.supercheck.io",
      "description": "Production (Cloud)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/jobs/run": {
      "post": {
        "operationId": "runJob",
        "summary": "Run a job immediately",
        "description": "Execute a job immediately, bypassing the cron schedule.",
        "tags": [
          "Jobs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "jobId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "jobId"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "runId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "CLI token (`sck_live_*`) for full API access, or trigger key (`sck_trigger_*`) for job trigger endpoint only."
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or token invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        }
      },
...
    },
...
  }
}